Added DB logging

+++ server/routes/wings.js
@@ -1,4 +1,5 @@
 const router = require("express").Router();
+const logger = require("../utils/logging");
 let Wing = require("../models/wing.model");
 
 router.get("/", async (req, res) => {
@@ -22,12 +23,15 @@ router.get("/:id", async (req, res) => {
 
 router.post("/", async (req, res) => {
   try {
+    if (!req.user) return res.status("403").json("Unauthenticated");
+
     const wing = new Wing({
       id: req.body.id,
       wingName: req.body.wingName,
       wingLabel: req.body.wingLabel,
     });
     const newWing = await wing.save();
+    logger.createLog(req.user.username, "added", "wing", newWing._id);
     res.json(newWing);
   } catch (err) {
     res.status(400).json(`Error: ${err}`);
@@ -36,12 +40,15 @@ router.post("/", async (req, res) => {
 
 router.put("/:id", async (req, res) => {
   try {
+    if (!req.user) return res.status("403").json("Unauthenticated");
+
     const wing = await Wing.findById(req.params.id);
     if (wing == null) return res.status(404).json("Not found.");
     wing.id = req.body.id;
     wing.wingName = req.body.wingName;
     wing.wingLabel = req.body.wingLabel;
     const updated = await wing.save();
+    logger.createLog(req.user.username, "updated", "wing", req.params.id);
     res.json(updated);
   } catch (err) {
     res.status(400).json(`Error: ${err}`);
@@ -50,8 +57,11 @@ router.put("/:id", async (req, res) => {
 
 router.delete("/:id", async (req, res) => {
   try {
+    if (!req.user) return res.status("403").json("Unauthenticated");
+
     const wing = await Wing.findByIdAndDelete(req.params.id);
     if (wing == null) return res.status(404).json("Not found.");
+    logger.createLog(req.user.username, "deleted", "wing", req.params.id);
     res.json(wing);
   } catch (err) {
     res.status(400).json(`Error ${err}`);
